Video를 DB에서 찾지 못했을 때 던져지는 예외를 수정
